接口自动化bug和细节修复

+++ src/main/java/com/rabbit/model/JobParams.java
@@ -9,13 +9,12 @@ import java.util.List;
 
 
 @Data
-public class JobParams implements Serializable  {
-
-	private static final long serialVersionUID = 1L;
-	private Long id;
-	private Integer envId;//环境id
-	private Long clientId;//客户端IP
-	private String  browserType;
-	private Integer isSendEmail;//执行完毕是否发送邮件
-	private String receivers ;//执行完毕是否发送邮件
+public class JobParams implements Serializable {
+    private static final long serialVersionUID = 1L;
+    private Long id;
+    private Long envId;//环境id
+    private Long clientId;//客户端IP
+    private String browserType;
+    private Integer isSendEmail;//执行完毕是否发送邮件
+    private String receivers;//执行完毕是否发送邮件
 }
